requirements.txt
imbalance-xgboost==0.8.1
imbalanced-databases==0.1.1
imblearn==0.0
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
joblib==1.0.0
lazypredict==0.2.9
Markdown==3.3.4
pandas-profiling==2.3.0
pep8==1.7.1
#pickle5==0.0.9
pickleshare==0.7.5
pipenv==2020.8.13
plotly-express==0.4.0
python-dateutil==2.8.1
scikit-image==0.15.0
scikit-learn==0.23.1
scipy==1.5.4
seaborn==0.9.0
sklearn==0.0
streamlit==0.84.1
streamlit-analytics==0.2.2
streamlit-metrics==0.1.0
streamlit-pandas-profiling==0.1.2
text-unidecode==1.2
unicodecsv==0.14.1
xgboost==1.1.1